pyproject.toml
[project]
authors = [{name = "cminalo", email = "cminalopez@outlook.com"}]
dependencies = [
    "loguru>=0.7.3",
    "python-dotenv>=1.0.1", "hatch>=1.16.3,<2", "psutil>=7.2.2,<8",
]
name = "chronos-logger"
requires-python = ">= 3.11"
version = "0.1.0"

[project.optional-dependencies]
dev = ["pytest"]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/chronos"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["osx-arm64"]

[tool.pixi.pypi-dependencies]
chronos-logger = { path = ".", editable = true, extras = ["dev"] }

[tool.pixi.tasks]

[tool.pixi.dependencies]
python = "3.13.*"
